Refactor express config fns, stub consent flow
* Refactor `lib/express` to be a dir, spcp and myinfo in diff files
* SPCP: If `esrvcID` is `MYINFO-CONSENTPLATFORM`, redirect to MockInfo
* Implement consent flow, but stop short at the actual consent screen

TODO: Figure out how to issue/use tokens for `/person`
