--- lib/express.js
@@ -1,152 +1,0 @@
-const _ = require('lodash')
-const bodyParser = require('body-parser')
-const fs = require('fs')
-const { pick, partition } = require('lodash')
-const morgan = require('morgan')
-const { render } = require('mustache')
-const path = require('path')
-const { DOMParser } = require('xmldom')
-const xpath = require('xpath')
-
-const assertions = require('./assertions')
-const crypto = require('./crypto')
-const samlArtifact = require('./saml-artifact')
-
-const domParser = new DOMParser()
-const dom = xmlString => domParser.parseFromString(xmlString)
-
-const TEMPLATE = fs.readFileSync(path.resolve(__dirname, '../static/saml/unsigned-response.xml'), 'utf8')
-const LOGIN_TEMPLATE = fs.readFileSync(path.resolve(__dirname, '../static/html/login-page.html'), 'utf8')
-
-function config (app, { showLoginPage, serviceProvider, idpConfig, port }) {
-  const { verifySignature, verifyMyInfoSignature, sign, promiseToEncryptAssertion } = crypto(serviceProvider)
-  app.use(morgan('combined'))
-
-  for (const idp of ['singPass', 'corpPass']) {
-    app.get(`/${idp.toLowerCase()}/logininitial`, (req, res) => {
-      const relayState = encodeURIComponent(req.query.Target)
-      if (showLoginPage) {
-        const identities = assertions.identities[idp]
-        const values = identities
-          .map((id, index) => {
-            const samlArt = samlArtifact(idpConfig[idp].id, index)
-            const assertURL =
-              `${idpConfig[idp].assertEndpoint}?SAMLart=${samlArt}&RelayState=${relayState}`
-            if (assertions.myinfo.personas[id]) {
-              id += ' [MyInfo]'
-            }
-            return { id, assertURL }
-          })
-        const response = render(LOGIN_TEMPLATE, values)
-        res.send(response)
-      } else {
-        const samlArt = samlArtifact(idpConfig[idp].id)
-        const assertURL =
-          `${idpConfig[idp].assertEndpoint}?SAMLart=${samlArt}&RelayState=${relayState}`
-        console.warn(`Redirecting login from ${req.query.PartnerId} to ${assertURL}`)
-        res.redirect(assertURL)
-      }
-    })
-
-    app.post(
-      `/${idp.toLowerCase()}/soap`,
-      bodyParser.text({ type: 'text/xml' }),
-      (req, res) => {
-        // Extract the body of the SOAP request
-        const { body } = req
-        const xml = dom(body)
-
-        if (!verifySignature(xml)) {
-          res.status(400).send('Request has bad signature')
-        } else {
-          // Grab the SAML artifact
-          // TODO: verify the SAML artifact is something we sent
-          // TODO: do something about the partner entity id
-          const samlArtifact = xpath.select("string(//*[local-name(.)='Artifact'])", xml)
-          console.warn(`Received SAML Artifact ${samlArtifact}`)
-          // Take the template and plug in the typical SingPass/CorpPass response
-          // Sign and encrypt the assertion
-          const samlArtifactBuffer = Buffer.from(samlArtifact, 'base64')
-          const index = samlArtifactBuffer.readInt8(samlArtifactBuffer.length - 1)
-          const assertion = assertions.identities[idp][index]
-            ? assertions[idp].create(assertions.identities[idp][index])
-            : assertions[idp].default
-          const signedAssertion = sign(assertion, "//*[local-name(.)='Assertion']")
-          promiseToEncryptAssertion(signedAssertion)
-            .then(assertion => {
-              const response = render(TEMPLATE, { assertion })
-              const signedResponse = sign(
-                sign(response, "//*[local-name(.)='Response']"),
-                "//*[local-name(.)='ArtifactResponse']"
-              )
-              res.send(signedResponse)
-            })
-        }
-      }
-    )
-  }
-
-  const lookupPerson = allowedAttributes => (req, res) => {
-    const requestedAttributes = (req.query.attributes || '').split(',')
-
-    const [attributes, disallowedAttributes] = partition(
-      requestedAttributes,
-      v => allowedAttributes.includes(v)
-    )
-
-    if (disallowedAttributes.length > 0) {
-      res.status(401).send({ code: 401, message: 'Disallowed', fields: disallowedAttributes.join(',') })
-    } else {
-      const persona = assertions.myinfo.personas[req.params.uinfin]
-      res.status(persona ? 200 : 404)
-        .send(
-          persona
-            ? pick(persona, attributes)
-            : { code: 404, message: 'UIN/FIN does not exist in MyInfo.', fields: '' }
-        )
-    }
-  }
-
-  const allowedAttributes = assertions.myinfo.attributes
-
-  app.get(
-    '/myinfo/person-basic/:uinfin/',
-    (req, res, next) => {
-      const [, authHeader] = req.get('Authorization').split(' ')
-      const authHeaderFieldPairs = _(authHeader)
-        .replace(/"/g, '')
-        .replace(/apex_l2_eg_/g, '')
-        .split(',')
-        .map(v => v.replace('=', '~').split('~'))
-
-      const authHeaderFields = _(authHeaderFieldPairs)
-        .fromPairs()
-        .mapKeys((v, k) => _.camelCase(k))
-        .value()
-
-      authHeaderFields.clientId = authHeaderFields.appId
-      authHeaderFields.singpassEserviceId = authHeaderFields.appId.replace(/^[^-]+-/, '')
-
-      authHeaderFields.httpMethod = req.method
-      // HACK: kludge the URL together
-      const hostPort = ['localhost', '127.0.0.1'].includes(req.hostname)
-        ? `${req.hostname}:${port}`
-        : req.hostname
-
-      authHeaderFields.url = `${req.protocol}://${hostPort}${req.baseUrl}${req.path}`
-      authHeaderFields.requestedAttributes = req.query.attributes
-
-      if (verifyMyInfoSignature(authHeaderFields.signature, authHeaderFields)) {
-        next()
-      } else {
-        res.status(403).send({ code: 403, message: 'Digital Service is invalid', fields: '' })
-      }
-    },
-    lookupPerson(allowedAttributes.basic)
-  )
-  app.get('/myinfo/person/:uinfin/', lookupPerson([...allowedAttributes.basic, ...allowedAttributes.income]))
-
-  return app
-}
-
-module.exports = { config }
