+++ lib/express/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  configSpcp: require('./spcp'),
+  configMyInfo: require('./myinfo'),
+}
